requirements.txt
albumentations==1.1.0
efficientnet_pytorch==0.7.1
lifelines==0.26.4
numpy==1.21.6
opencv_contrib_python==4.8.0.74
opencv_python==4.4.0.46
opencv_python_headless==4.2.0.34
openslide_python==1.2.0
pandas==1.3.5
Pillow==8.4.0
rich==10.16.1
scikit_image==0.19.1
scikit_learn==1.0.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu110
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.2+cu110
tqdm==4.62.3
yacs==0.1.8
h5py==2.9.0
numba==0.44
torchtuples==0.2.0